+++ src/JSON.ts
@@ -2,6 +2,7 @@
  * @since 0.1.0
  */
 
+import { Newtype, iso } from 'newtype-ts';
 import { Either } from 'fp-ts/Either';
 import * as E from 'fp-ts/Either';
 import { Option } from 'fp-ts/Option';
@@ -10,15 +11,25 @@ import { flow, identity, pipe } from 'fp-ts/function';
 import { isString } from './String';
 
 /**
+ * Newtype representing stringified JSON.
+ *
+ * @since 0.5.0
+ */
+export type JSONString = Newtype<{ readonly JSONString: unique symbol }, string>;
+
+const { wrap: mkJSONString, unwrap: unJSONString } = iso<JSONString>();
+
+/**
  * Stringify some arbitrary data.
  *
  * @since 0.1.0
  */
-export const stringify = <E>(f: (e: TypeError) => E) => (x: unknown): Either<E, string> => pipe(
+export const stringify = <E>(f: (e: TypeError) => E) => (x: unknown): Either<E, JSONString> => pipe(
     // It should only throw some sort of `TypeError`:
     // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify
     E.stringifyJSON(x, (e) => f(e as TypeError)),
     E.filterOrElse(isString, () => f(new TypeError('Stringify output not a string'))),
+    E.map(mkJSONString),
 );
 
 /**
@@ -26,7 +37,7 @@ export const stringify = <E>(f: (e: TypeError) => E) => (x: unknown): Either<E, 
  *
  * @since 0.1.0
  */
-export const stringifyO: (data: unknown) => Option<string> =
+export const stringifyO: (data: unknown) => Option<JSONString> =
     flow(stringify(identity), O.fromEither);
 
 /**
@@ -34,8 +45,17 @@ export const stringifyO: (data: unknown) => Option<string> =
  *
  * @since 0.1.0
  */
-export const stringifyPrimitive = (x: string | number | boolean | null): string =>
-    JSON.stringify(x);
+export const stringifyPrimitive = (x: string | number | boolean | null): JSONString =>
+    pipe(x, JSON.stringify, mkJSONString);
+
+/**
+ * Parse a string as JSON. This is safe provided there have been no shenanigans
+ * with the `JSONString` newtype.
+ *
+ * @since 0.5.0
+ */
+export const unstringify: (x: JSONString) => unknown =
+    flow(unJSONString, JSON.parse);
 
 /**
  * Parse a string as JSON.