+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to semantic versioning.
 - Add `JSONString` newtype to the `JSON` module, and update various functions in this module accordingly. This adds `newtype-ts` and `monocle-ts` as peer dependencies.
 - Add `insertMany` function to the `Array` module.
 - Add `withIndex` function to the `Function` module.
+- Add `elapsed` function to the `Task` module.
 - Add `both` and `either` functions to the `Boolean` module.
 - Add `unsafeUnwrapLeft` function to the `Either` module.
 - Replace `exec` in the `String` module with `matchAll`.